refactor(612): extract VariationProvenanceZones from ReviewFormFields
Verbatim extraction so the three-zone picker can be mounted on every curation
surface, not just the Review page's edit modal. Every data-testid is preserved
character-for-character and ReviewFormFields.provenance.spec.ts passes unchanged,
which is the proof.

displayName is a prop rather than logic in the component: the Review form's
resolver also consults its option-tree labels, which no other surface has.
