+++ app/src/views/curate/components/ReviewFormFields.vue
@@ -82,97 +82,11 @@
         pre-#608 plain picker — which is production on day one, before the
         provenance backfill runs.
       -->
-      <div v-if="zonesActive" class="vario-zones" data-testid="variation-provenance-zones">
-        <section
-          v-if="zoneConfirmed.length"
-          class="vario-zones__zone"
-          aria-labelledby="vario-zone-confirmed-heading"
-          data-testid="variation-zone-confirmed"
-        >
-          <h6 id="vario-zone-confirmed-heading" class="vario-zones__heading">
-            Confirmed
-            <span class="vario-zones__count" data-testid="variation-zone-confirmed-count">
-              {{ termCount(zoneConfirmed.length) }}
-            </span>
-          </h6>
-          <ul class="vario-zones__chips">
-            <li
-              v-for="entry in zoneConfirmed"
-              :key="entry.tag"
-              data-testid="variation-confirmed-chip"
-              :data-tag="entry.tag"
-            >
-              <span class="sysndd-chip sysndd-chip--neutral">
-                <i class="bi bi-check2 me-1" aria-hidden="true" />
-                {{ displayName(entry) }} ({{ entry.modifierLabel }})
-              </span>
-            </li>
-          </ul>
-        </section>
-
-        <section
-          v-if="zoneNeedsConfirmation.length"
-          class="vario-zones__zone"
-          aria-labelledby="vario-zone-needs-confirmation-heading"
-          data-testid="variation-zone-needs-confirmation"
-        >
-          <h6 id="vario-zone-needs-confirmation-heading" class="vario-zones__heading">
-            Needs confirmation
-            <span class="vario-zones__count" data-testid="variation-zone-needs-confirmation-count">
-              {{ termCount(zoneNeedsConfirmation.length) }}
-            </span>
-          </h6>
-          <p class="vario-zones__hint">
-            Machine-derived from the sources below. Saving leaves an unconfirmed term in place; only
-            Confirm records that you agree with it.
-          </p>
-          <ul class="vario-zones__cards">
-            <VariationProvenanceCard
-              v-for="entry in zoneNeedsConfirmation"
-              :key="entry.tag"
-              :entry="entry"
-              :display-name="displayName(entry)"
-              primary-label="Confirm"
-              secondary-label="Remove"
-              primary-testid="variation-action-confirm"
-              secondary-testid="variation-action-remove"
-              :disabled="readonly"
-              @primary="variationZones?.confirmTerm(entry.tag)"
-              @secondary="variationZones?.removeTerm(entry.tag)"
-            />
-          </ul>
-        </section>
-
-        <section
-          v-if="zoneSuggested.length"
-          class="vario-zones__zone"
-          aria-labelledby="vario-zone-suggested-heading"
-          data-testid="variation-zone-suggested"
-        >
-          <h6 id="vario-zone-suggested-heading" class="vario-zones__heading">
-            Suggested — not in the entity
-            <span class="vario-zones__count" data-testid="variation-zone-suggested-count">
-              {{ termCount(zoneSuggested.length) }}
-            </span>
-          </h6>
-          <p class="vario-zones__hint">Candidate terms. Nothing is added until you accept it.</p>
-          <ul class="vario-zones__cards">
-            <VariationProvenanceCard
-              v-for="entry in zoneSuggested"
-              :key="entry.tag"
-              :entry="entry"
-              :display-name="displayName(entry)"
-              primary-label="Accept"
-              secondary-label="Dismiss"
-              primary-testid="variation-action-accept"
-              secondary-testid="variation-action-dismiss"
-              :disabled="readonly"
-              @primary="variationZones?.acceptSuggestion(entry.tag)"
-              @secondary="variationZones?.dismissSuggestion(entry.tag)"
-            />
-          </ul>
-        </section>
-      </div>
+      <VariationProvenanceZones
+        :variation-zones="variationZones"
+        :readonly="readonly"
+        :display-name="displayName"
+      />
 
       <TreeMultiSelect
         v-if="variationOptions && variationOptions.length > 0"
@@ -332,7 +246,7 @@
 <script setup lang="ts">
 import { computed } from 'vue';
 import TreeMultiSelect from '@/components/forms/TreeMultiSelect.vue';
-import VariationProvenanceCard from '@/views/curate/components/VariationProvenanceCard.vue';
+import VariationProvenanceZones from '@/views/curate/components/VariationProvenanceZones.vue';
 import type { ReviewFormData } from '@/views/curate/composables/useReviewForm';
 import type {
   VariationProvenanceZonesApi,
@@ -382,10 +296,6 @@ const localFormData = computed({
 
 // The zones API exposes refs, and a prop object's nested refs are NOT unwrapped
 // in the template, so each zone is re-exposed through a local computed.
-const zonesActive = computed(() => Boolean(props.variationZones?.hasZones.value));
-const zoneConfirmed = computed(() => props.variationZones?.confirmed.value ?? []);
-const zoneNeedsConfirmation = computed(() => props.variationZones?.needsConfirmation.value ?? []);
-const zoneSuggested = computed(() => props.variationZones?.suggested.value ?? []);
 
 /**
  * Flatten the TreeMultiSelect options so a term the API did not name (e.g. one
@@ -406,15 +316,19 @@ const variationLabelById = computed(() => {
   return labels;
 });
 
-/** Server-supplied name wins; then the tree label; then the bare CURIE. */
+/**
+ * Server-supplied name wins; then the tree label; then the bare CURIE.
+ *
+ * Passed INTO VariationProvenanceZones rather than living there: the tree label
+ * comes from `variationOptions`, which only this form has. The other surfaces
+ * fall back to the component's own name-or-CURIE default rather than showing a
+ * label nobody supplied.
+ */
 const displayName = (entry: VariationZoneEntry): string =>
   entry.varioName ||
   variationLabelById.value.get(entry.tag) ||
   variationLabelById.value.get(entry.varioId) ||
   entry.varioId;
-
-/** Honest, grammatical count — exposed in TEXT, not conveyed by position. */
-const termCount = (count: number): string => `${count} term${count === 1 ? '' : 's'}`;
 
 /**
  * PMID tag validator
@@ -436,60 +350,6 @@ const tagValidatorPMID = (tag: string): boolean => {
    Quiet by design: one bounded surface, neutral tokens, no status hues. An
    unconfirmed annotation is un-reviewed, not broken — saturated colour is
    reserved for the action buttons inside VariationProvenanceCard. */
-.vario-zones {
-  display: grid;
-  gap: 0.6rem;
-  margin: 0.35rem 0 0.5rem;
-  padding: 0.6rem;
-  border: 1px solid var(--border-subtle, #d9e0ea);
-  border-radius: var(--radius-md, 0.375rem);
-  background: var(--neutral-50, #fafafa);
-}
-
-.vario-zones__zone {
-  display: grid;
-  gap: 0.35rem;
-  min-width: 0;
-}
-
-.vario-zones__heading {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: baseline;
-  justify-content: space-between;
-  gap: 0.5rem;
-  margin: 0;
-  color: var(--neutral-900, #212121);
-  font-size: 0.75rem;
-  font-weight: 700;
-  letter-spacing: 0.02em;
-  text-transform: uppercase;
-}
-
-.vario-zones__count {
-  color: var(--neutral-600, #757575);
-  font-size: 0.75rem;
-  font-weight: 600;
-  text-transform: none;
-  letter-spacing: normal;
-}
-
-.vario-zones__hint {
-  margin: 0;
-  color: var(--neutral-600, #757575);
-  font-size: 0.75rem;
-  line-height: 1.4;
-}
-
-.vario-zones__chips,
-.vario-zones__cards {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.35rem;
-  margin: 0;
-  padding: 0;
-  list-style: none;
-}
 
 .vario-zones__cards {
   display: grid;
